requirements.txt
pandas==2.2.2
openpyxl==3.1.5
confluent-kafka==2.5.3
neo4j==5.24.0
python-dotenv==1.0.1
pydantic==2.8.2